chore(userspace/libsinsp)!: pass `targetbuf_size` as `size_t`
Update `sinsp_utils::sockinfo_to_str()` to accept a `targetbuf_size`
as a `size_t` parameter.

BREAKING CHANGE: change `targetbuf_size` type

+++ userspace/libsinsp/utils.h
@@ -78,7 +78,7 @@ public:
 	static bool sockinfo_to_str(sinsp_sockinfo* sinfo,
 	                            scap_fd_type stype,
 	                            char* targetbuf,
-	                            uint32_t targetbuf_size,
+	                            size_t targetbuf_size,
 	                            bool resolve = false);
 
 	//
